feat: modular mission map for game dashboard
I did modularize the mission map on web_console/game_dashboard to unify avatar streams and chakra metrics, expecting operators to navigate large buttons with keyboards or gamepads.
