+++ web_console/game_dashboard/index.html
@@ -5,7 +5,7 @@
   <title>Game Dashboard</title>
   <style>
     body { font-family: Arial, sans-serif; text-align: center; margin: 2rem; }
-    .mission-map button { font-size: 1.5rem; margin: 1rem; padding: 1rem 2rem; }
+    .mission-map button { font-size: 2rem; margin: 1rem; padding: 1.25rem 2.5rem; }
     .modal-overlay { position: fixed; top: 0; left: 0; width: 100%; height: 100%; background: rgba(0,0,0,0.6); display: flex; align-items: center; justify-content: center; }
     .modal { background: #fff; padding: 1.5rem; border-radius: 8px; max-width: 420px; width: 100%; text-align: left; }
     .progress { height: 8px; background: #ddd; border-radius: 4px; overflow: hidden; margin-bottom: 1rem; }
